Moving some common controller and model features into concerns to dry up the code.  Big refucktor cum'ing up in another branch :(
