+++ app/models/application_record.rb
@@ -2,5 +2,7 @@
 
 # All models inherit from this class.
 class ApplicationRecord < ActiveRecord::Base
+  include ScopeById
+
   self.abstract_class = true
 end
