+++ app/models/role.rb
@@ -31,4 +31,7 @@ class Role < ApplicationRecord
   # --------------------------------------------------------------------------------------------------------------------
 
   has_and_belongs_to_many :users
+
+  # Scopes
+  # --------------------------------------------------------------------------------------------------------------------
 end
