+++ app/models/session_activity.rb
@@ -23,4 +23,7 @@ class SessionActivity < ApplicationRecord
   # --------------------------------------------------------------------------------------------------------------------
 
   belongs_to :session
+
+  # Scopes
+  # --------------------------------------------------------------------------------------------------------------------
 end
