+++ app/models/user.rb
@@ -47,8 +47,6 @@ class User < ApplicationRecord
   # Scopes
   # --------------------------------------------------------------------------------------------------------------------
 
-  scope :by_id, ->(ids) { where id: ids }
-
   # Pundit Helpers
   # --------------------------------------------------------------------------------------------------------------------
